chore(component): move 'pageLinks.tsx' to src/component/Drawer folder
